refactor: decompose reports-manager.tsx into focused sub-components (#142)
Break down the 1228-line monolithic ReportsManager into 7 focused modules:
- types.ts: shared interfaces and constants
- use-report-form.ts: custom hook for form state management
- report-form.tsx: report create/edit form
- report-card.tsx: individual report card with actions
- generated-reports-list.tsx: generated reports history
- json-result-preview.tsx: JSON report result with summary cards
- reports-manager.tsx: slim orchestrator (353 lines)

Original file re-exports for backward compatibility.
