+++ src/components/reports/index.ts
@@ -1,0 +1,2 @@
+export { ReportsManager } from "./reports-manager";
+export type { Account, Category, ReportFilters, SavedReport, GeneratedReport, JsonReportResult } from "./types";
